refactor: implement pessimistic locking, fix N+1 queries, and harden validation
Concurrency & Database Locking:
- Added `@Version` to `Booking` entity to enable JPA optimistic locking.
- Introduced `findByIdAndLock` with `@Lock(LockModeType.PESSIMISTIC_WRITE)` in `PaymentRepository` to prevent race conditions during parallel Stripe webhook deliveries and refund processing.
- Swapped `findAllById` to `findAllByIdWithLock` in `SeatService` to guarantee database-level pessimistic locks accompany the Redis distributed locks during checkout.

Performance Optimization (N+1 Fix):
- Created `EventSeatCount` JPA projection interface.
- Implemented `countAvailableSeatsByEventIds` in `SeatRepository` using a `GROUP BY` clause for batch fetching.
- Refactored `EventService.buildEventResponsePage` to fetch all seat counts for a page of events in a single round-trip, eliminating the N+1 query bottleneck.

Data Types & Validation:
- Migrated all seat counting fields (`numberOfSeats`, `availableSeats`) from `Integer` to `Long` across entities, DTOs, and services.
- Added `@Size(min = 1)` validation to `HoldSeatsRequest` to prevent empty arrays.
- Added explicit null and zero-amount validation in `PaymentService.initializePayment` to throw `InvalidAmountException`.

Resilience & Stability:
- Added defensive null checks to `StripeWebhookController` metadata extraction and `SeatService` Redis lock releases (preventing NPEs on orphaned seats).
- Wrapped `EventCancellationListener` in a robust try-catch block to ensure exceptions during mass-refunds are logged and re-thrown cleanly.
- Added extensive debug logging to `BookingReferenceGenerator` retries.

Documentation & Testing:
- Committed the enterprise-grade Javadocs for all domain Enums (BookingStatus, EventStatus, PaymentMethod, etc.).
- Updated `PaymentServiceTest`, `SeatServiceTest`, and `EventControllerTest` to mock the new locking repository methods and assert the new `Long` data types and validation rules.
- Updated E2E `BookingFlowIntegrationTest` payloads to reflect the `Long` seat counts.

+++ src/main/java/io/github/kxng0109/quicktix/dto/request/CreateEventRequest.java
@@ -78,7 +78,7 @@ public record CreateEventRequest(
 		)
 		@NotNull(message = "Number of seats can't be blank")
 		@PositiveOrZero(message = "Number of seats must be 0 or greater.")
-		Integer numberOfSeats
+		Long numberOfSeats
 ) {
 	@AssertTrue(message = "Start date time must be before end date time")
 	@Schema(hidden = true)
